Remove loadMessages and load and simplify estructuring logic
