+++ src/api/index.js
@@ -4,5 +4,4 @@ export * from './getArticleSummary'
 export * from './getLanglinks'
 export * from './getRandomArticleTitle'
 export * from './getSuggestedArticles'
-export * from './language'
 export * from './search'
