+++ src/index.js
@@ -1,15 +1,14 @@
 import 'preact/debug'
 import { h, render } from 'preact'
 import Banana from 'banana-i18n'
-import { loadAllLanguagesMessages } from 'api'
+import { getAppLanguage, loadAllLanguagesMessages } from 'utils'
 import { App } from 'components'
-import { getAppLanguage } from 'utils'
+
 import '../style/style.less'
 
 const lang = getAppLanguage() || navigator.language.substr(0, 2)
 const banana = new Banana(lang)
 
-const allMessages = loadAllLanguagesMessages()
-banana.load(allMessages)
+banana.load(loadAllLanguagesMessages())
 
 render(<App i18n={banana} />, document.querySelector('.root'))
